High scores loaded asynchronously

+++ src/taxomania/games/mastermind/HighScoreActivity.java
@@ -6,6 +6,7 @@ import java.util.Map;
 import android.app.Activity;
 import android.content.pm.ActivityInfo;
 import android.graphics.Color;
+import android.os.AsyncTask;
 import android.os.Bundle;
 import android.view.Gravity;
 import android.view.Menu;
@@ -24,7 +25,7 @@ public class HighScoreActivity extends Activity {
         highScoreTable = (TableLayout) findViewById(R.id.highScoreTable);
         setRequestedOrientation(ActivityInfo.SCREEN_ORIENTATION_PORTRAIT);
 
-        printHighscores();
+        new PrintHighScores().execute(false);
     }
 
     private void addHeaders() {
@@ -54,27 +55,20 @@ public class HighScoreActivity extends Activity {
         tr.addView(gText);
     }
 
-    private void printHighscores() {
-        highScoreTable.removeAllViewsInLayout();
-        addHeaders();
-
-
-        final DataHelper dh = DataHelper.getInstance(this);
-        List<Map<String, Object>> scores = dh.selectAll();
-
+    private void printScores(final List<Map<String, Object>> scores){
         int i = 1;
         for (Map<String, Object> score : scores) {
-            TableRow tr = new TableRow(this);
+            final TableRow tr = new TableRow(this);
             highScoreTable.addView(tr);
 
-            TextView rankText = new TextView(this);
-            rankText.setText(((Integer)i).toString());
+            final TextView rankText = new TextView(this);
+            rankText.setText(((Integer) i).toString());
             rankText.setTextColor(Color.WHITE);
             rankText.setGravity(Gravity.CENTER);
             tr.addView(rankText);
             i++;
 
-            TextView nameText = new TextView(this);
+            final TextView nameText = new TextView(this);
             nameText.setText(score.get(DataHelper.Scores.NAME).toString());
             nameText.setTextColor(Color.WHITE);
             tr.addView(nameText);
@@ -82,45 +76,65 @@ public class HighScoreActivity extends Activity {
             final int time = (Integer) score.get(DataHelper.Scores.TIME);
             final int minutes = time / 60;
             final int seconds = time % 60;
-            final String secs = (seconds < 10) ? ("0" + seconds) : ((Integer) seconds).toString();
-            final String mins = (minutes < 10) ? ("0" + minutes) : ((Integer) minutes).toString();
+            final String secs = (seconds < 10) ? ("0" + seconds) : ((Integer) seconds)
+                    .toString();
+            final String mins = (minutes < 10) ? ("0" + minutes) : ((Integer) minutes)
+                    .toString();
 
-            TextView scoreText = new TextView(this);
+            final TextView scoreText = new TextView(this);
             scoreText.setText(mins + ":" + secs);
             scoreText.setTextColor(Color.WHITE);
             tr.addView(scoreText);
 
-            TextView gText = new TextView(this);
+            final TextView gText = new TextView(this);
             gText.setText(score.get(DataHelper.Scores.GUESSES).toString());
             gText.setGravity(Gravity.CENTER);
             gText.setTextColor(Color.WHITE);
             tr.addView(gText);
-        }
-    }
+        } // for each score
+    } // printScores
+
+    private final class PrintHighScores extends AsyncTask<Boolean, Void, List<Map<String, Object>>>{
+
+        @Override
+        protected void onPreExecute() {
+            highScoreTable.removeAllViewsInLayout();
+            addHeaders();
+        } // onPreExecute
+
+        @Override
+        protected List<Map<String, Object>> doInBackground(Boolean... delete) {
+            final DataHelper dh = DataHelper.getInstance(HighScoreActivity.this);
+            if (delete[0]){
+                dh.deleteAll();
+            }
+            return dh.selectAll();
+        } // doInBackground
+
+        @Override
+        protected void onPostExecute(final List<Map<String, Object>> scores) {
+            printScores(scores);
+        } // onPostExecute
+    } // PrintHighScores
 
     private static final int MENU_RESET = Menu.FIRST;
+
     @Override
-    public boolean onCreateOptionsMenu(final Menu menu)
-    {
+    public boolean onCreateOptionsMenu(final Menu menu) {
         super.onCreateOptionsMenu(menu);
         menu.add(0, MENU_RESET, 0, "Reset High Scores");
         return true;
     } // onCreateOptionsMenu
 
     @Override
-    public boolean onOptionsItemSelected(final MenuItem item)
-    {
-        switch (item.getItemId())
-        {
+    public boolean onOptionsItemSelected(final MenuItem item) {
+        switch (item.getItemId()){
             case MENU_RESET:
-                final DataHelper dh = DataHelper.getInstance(this);
-                dh.deleteAll();
-                printHighscores();
+                new PrintHighScores().execute(true);
                 return true;
             default:
                 return super.onOptionsItemSelected(item);
         } // switch
     } // onOptionsItemSelected
-
 
 } // class HighScoreActivity
